+++ qsiprep/workflows/fieldmap/fmap.py
@@ -22,7 +22,7 @@ of the BIDS specification.
 
 from nipype.pipeline import engine as pe
 from nipype.interfaces import utility as niu, fsl, ants
-from nipype.workflows.dmri.fsl.utils import demean_image, cleanup_edge_pipeline
+from .utils import demean_image, cleanup_edge_pipeline
 from ...niworkflows.engine.workflows import LiterateWorkflow as Workflow
 from ...niworkflows.interfaces.images import IntraModalMerge
 from ...niworkflows.interfaces.masks import BETRPT
